+++ tests/Fixtures/CustomException.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace MF\Validator\Fixtures;
+
+class CustomException extends \Exception
+{
+}