+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Added, Changed, Fixed, Removed) should be added as needed. -->
 ## Unreleased
+- Allow custom `Exception` to be thrown on assertion failed
 
 ## 2.3.0 - 2018-10-16
 - [dev] update and fix code style